1.0 Marketing website
First commit for the website. This page is for selling our product to customers. I will begin adding the log in features and profile features in the top right corner of the page for our users. Note: after each commit I will increment the 1.0 to keep track of how many updates we have made to the original file.